refactor: 멀티모듈로 구조 수정
