+++ project/dayoung-onboarding/onboarding-cloud/build.gradle.kts
@@ -1,3 +1,11 @@
 dependencies {
     testImplementation("org.springframework.boot:spring-boot-starter-test")
 }
+
+tasks.getByName<org.springframework.boot.gradle.tasks.bundling.BootJar>("bootJar") {
+    enabled = false
+}
+
+tasks.getByName<Jar>("jar") {
+    enabled = true
+}
