+++ project/dayoung-onboarding/onboarding-data/build.gradle.kts
@@ -1,6 +1,4 @@
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("com.h2database:h2")
-
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
 }
